data_ingestion/requirements.txt
requests==2.31.0
nltk==3.8.1
dacite==1.8.1
pyyaml==6.0.1
opensearch-py==2.6.0
